--- src/lib/causalTypeColors.tsx
@@ -1,12 +1,0 @@
-/**
- * Maps causal type strings to hex colors, for styling
- * diagram elements and dependency arrows.
- */
-export const causalTypeColors: Record<string, string> = {
-    Lever: "#3946fae5",
-    Intermediate: "#0c9188e5",
-    Outcome: "#ad3e9de5",
-    External: "#125c26e5",
-    Unknown: "#7a7a7ae5",
-    Selected: "#ffffff",
-}
